+++ apps/mis-server/src/services/init.ts
@@ -140,7 +140,11 @@ export const initServiceServer = plugin((server) => {
     },
 
     completeInit: async ({ em }) => {
+
+
       const initializationTime = new SystemState(SystemState.KEYS.INITIALIZATION_TIME, new Date().toISOString());
+
+
 
       try {
         await em.persistAndFlush(initializationTime);
