+++ apps/mis-web/src/apis/api.mock.ts
@@ -107,6 +107,10 @@ const mockUsers = [
 
 export const mockApi: MockApi<typeof api> = {
 
+  getMissingDefaultPriceItems: async () => {
+    return { items: ["test.test", "test1.test2"]};
+  },
+
   getAllTenants: async () => (
     {
       totalCount: 3,
